Fix eslint, rename file
Fixed ESlint warning by nesting children instead of passing as props. Renamed file StoryStarterComponent to StoryStarterPage to make the name distinct from the components folder.
